+++ Controllers/UsersController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using BlogVer2;
 using BlogVer2.Data;
+using Microsoft.AspNetCore.Http;
 
 namespace BlogVer2.Controllers
 {
@@ -22,12 +23,20 @@ namespace BlogVer2.Controllers
         // GET: Users
         public async Task<IActionResult> Index()
         {
+            if (HttpContext.Session.GetString("user") == null)
+            {
+                return RedirectToAction("Login", "Users");
+            }
             return View(await _context.User.ToListAsync());
         }
 
         // GET: Users/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (HttpContext.Session.GetString("user") == null)
+            {
+                return RedirectToAction("Login", "Users");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -46,6 +55,10 @@ namespace BlogVer2.Controllers
         // GET: Users/Create
         public IActionResult Create()
         {
+            if (HttpContext.Session.GetString("user") == null)
+            {
+                return RedirectToAction("Login", "Users");
+            }
             return View();
         }
 
@@ -56,6 +69,10 @@ namespace BlogVer2.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name,Password")] User user)
         {
+            if (HttpContext.Session.GetString("user") == null)
+            {
+                return RedirectToAction("Login", "Users");
+            }
             if (ModelState.IsValid)
             {
                 _context.Add(user);
@@ -68,6 +85,10 @@ namespace BlogVer2.Controllers
         // GET: Users/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (HttpContext.Session.GetString("user") == null)
+            {
+                return RedirectToAction("Login", "Users");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -88,6 +109,10 @@ namespace BlogVer2.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Password")] User user)
         {
+            if (HttpContext.Session.GetString("user") == null)
+            {
+                return RedirectToAction("Login", "Users");
+            }
             if (id != user.Id)
             {
                 return NotFound();
@@ -119,6 +144,10 @@ namespace BlogVer2.Controllers
         // GET: Users/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (HttpContext.Session.GetString("user") == null)
+            {
+                return RedirectToAction("Login", "Users");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -139,6 +168,10 @@ namespace BlogVer2.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (HttpContext.Session.GetString("user") == null)
+            {
+                return RedirectToAction("Login", "Users");
+            }
             var user = await _context.User.FindAsync(id);
             _context.User.Remove(user);
             await _context.SaveChangesAsync();
@@ -149,5 +182,45 @@ namespace BlogVer2.Controllers
         {
             return _context.User.Any(e => e.Id == id);
         }
+
+
+
+
+        //////////////////////////
+        // GET: Users/Create
+        public IActionResult Login()
+        {
+            return View();
+        }
+
+        // POST: Users/Create
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Login([Bind("Id,Name,Password")] User user)
+        {
+            if (ModelState.IsValid)
+            {
+                var validuser = from date in _context.User where (date.Name == user.Name && date.Password == user.Password) select date;
+                if (validuser.Count() > 0)
+                {
+                HttpContext.Session.SetString("user",user.Name);
+                return RedirectToAction("index","Home");
+                }
+            }
+            return RedirectToAction(nameof(Login));
+        }
+        ////////////////////////
+
+
+        //////////////////////////
+        // GET: Users/Create
+        public IActionResult LogOut()
+        {
+            HttpContext.Session.Remove("user");
+            return View();
+        }
+
     }
 }
